Recaptcha Emanled in login

+++ SpringProject/src/main/java/com/shreejit/springProject/Controller/EmployeeController.java
@@ -1,8 +1,11 @@
 package com.shreejit.springProject.Controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,47 +17,72 @@ import com.shreejit.springProject.model.Employee;
 @Controller
 public class EmployeeController {
 
-	@Autowired  
+	@Autowired
 	private EmployeeDao eDao;
 
 	@RequestMapping(value = "/employee", method = RequestMethod.GET)
-	public String getEmployeeForm(Model model) {
-		model.addAttribute("emodel",new Employee());
+	public String getEmployeeForm(Model model, HttpSession session) {
+
+		if (StringUtils.isEmpty(session.getAttribute("activeUser"))) {
+			return "login";
+		}
+
+		model.addAttribute("emodel", new Employee());
 		return "employeeForm";
 	}
-	
-	
+
 	@RequestMapping(value = "/employee", method = RequestMethod.POST)
-	public String saveEmployee(@ModelAttribute Employee emp, Model model) {
-		
-			eDao.addEmployee(emp);
-			model.addAttribute("elist", eDao.getAllEmployee());
-			
-			return "home";
+	public String saveEmployee(@ModelAttribute Employee emp, Model model, HttpSession session) {
+
+		if (StringUtils.isEmpty(session.getAttribute("activeUser"))) {
+			return "login";
 		}
-	
+
+		eDao.addEmployee(emp);
+		model.addAttribute("elist", eDao.getAllEmployee());
+
+		return "home";
+	}
+
 	@RequestMapping(value = "/{id}/delete", method = RequestMethod.GET)
-	public String deleteEmployee(@PathVariable("id") int id, Model model) {
-		
+	public String deleteEmployee(@PathVariable("id") int id, Model model,HttpSession session) {
+
+		if(StringUtils.isEmpty(session.getAttribute("activeUser"))) {
+			return "login";
+		}	
 		eDao.deleteEmployee(id);
 		model.addAttribute("elist", eDao.getAllEmployee());
 		return "home";
 	}
 
 	@RequestMapping(value = "/{id}/edit", method = RequestMethod.GET)
-	public String editEmployee(@PathVariable("id") int id, Model model) {
-		
-		model.addAttribute("emodel",eDao.getById(id));
+	public String editEmployee(@PathVariable("id") int id, Model model,HttpSession session) {
+
+		if(StringUtils.isEmpty(session.getAttribute("activeUser"))) {
+			return "login";
+		}
+		model.addAttribute("emodel", eDao.getById(id));
 		return "editForm";
 	}
-	
+
 	@RequestMapping(value = "/update", method = RequestMethod.POST)
-	public String editEmployee(@ModelAttribute Employee emp,Model model) {
-		
+	public String editEmployee(@ModelAttribute Employee emp, Model model,HttpSession session) {
+		if(StringUtils.isEmpty(session.getAttribute("activeUser"))) {
+			return "login";
+		}
 		eDao.updateEmployee(emp);
 		model.addAttribute("elist", eDao.getAllEmployee());
 		return "home";
 	}
-	
+
+	@RequestMapping(value = "/home", method = RequestMethod.GET)
+	public String home(Model model,HttpSession session) {
+		if(StringUtils.isEmpty(session.getAttribute("activeUser"))) {
+			return "login";
+		}
+		model.addAttribute("elist", eDao.getAllEmployee());
+		return "home";
+
 	}
 
+}
